% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_quantities.R
\name{calc_Astar}
\alias{calc_Astar}
\title{Calculate the altered genetic related matrix of Hunter et al. (2015)}
\usage{
calc_Astar(ped, cohort)
}
\arguments{
\item{ped}{a pedigree}

\item{cohort}{the cohort (containing one or more individuals) used to create the truncated pedigree}
}
\value{
a matrix containing the additive genetic relatedness value between all individuals calculated based on the truncated pedigree
}
\description{
Calculate the altered genetic related matrix of Hunter et al. (2015)
}
