% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{calc_completeness}
\alias{calc_completeness}
\title{Calculate completeness for individuals in the pedigree}
\usage{
calc_completeness(ped, id = "all", max_gen = "all")
}
\arguments{
\item{ped}{A pedigree dataframe in the format output by the process_ped function.}

\item{id}{The id of the individual(s) whose completeness you want calculated. The option 'all' can be used if you want the completeness of all individuals to be calculated.}

\item{max_gen}{For each individual, the maximum number of generations for which to calculate completeness. The option 'all' will result in completeness being calculated for all generations present in the generation for each individual.}
}
\value{
A dataframe containing completeness information for individuals in the pedigree.
}
\description{
Calculate completeness for individuals in the pedigree
}
