% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_quantities.R
\name{calc_gen_contr}
\alias{calc_gen_contr}
\title{Calculate the expected genetic contribution of a set of individuals to another set based on the Hunter et al. (2019) method}
\usage{
calc_gen_contr(
  ped,
  contributors,
  recipients,
  focal_contemporaries = NULL,
  standardize = c("none", "one", "two")
)
}
\arguments{
\item{ped}{pedigree dataframe}

\item{contributors}{the cohort for which to calculate genetic contribution}

\item{recipients}{the "recipient" group for which to calculate the genetic contribution of the individual(s) included in contributors}

\item{focal_contemporaries}{the individuals alive at the same time as the contributor cohort. This is only necessary if standardize is set to two.}

\item{standardize}{the method used to standardize the genetic contribution values.
This includes:
none: unstandardized value is returned
one: standardize the contribution value by the size of the recipients cohort
two: standardize the contribution value by the expected contribution value of all individuals alive at the same time as the contributors cohort. For this standardization, the ids of individuals in the population that exists at the same time as the contributors cohort must be input using the focal_contemporaries argument}
}
\value{
the expected genetic contribution of the contributors cohort to the recipients cohort
}
\description{
Calculate the expected genetic contribution of a set of individuals to another set based on the Hunter et al. (2019) method
}
