% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbr_coanc.R
\name{calc_kinmat}
\alias{calc_kinmat}
\title{Calculate kinship/coancestry matrix from a pedigree}
\usage{
calc_kinmat(ped, id_col = 1, sire_col = 2, dam_col = 3)
}
\arguments{
\item{ped}{the pedigree from which to calculate the kinship matrix}

\item{id_col}{the name or index of the id column}

\item{sire_col}{the name or index of the sire column}

\item{dam_col}{the name or index of the dam column}
}
\value{
the kinship matrix
}
\description{
Calculate kinship/coancestry matrix from a pedigree
}
