% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_quantities.R
\name{calc_partial_kinmat}
\alias{calc_partial_kinmat}
\title{Calculate partial kinship matrix for a given pedigree and focal founder}
\usage{
calc_partial_kinmat(
  ped,
  id_col = 1,
  sire_col = 2,
  dam_col = 3,
  founder_val = 0,
  focal_founder
)
}
\arguments{
\item{ped}{the pedigree from which to calculate the partial kinship matrix}

\item{id_col}{the name or index of the id column}

\item{sire_col}{the name or index of the sire column}

\item{dam_col}{the name or index of the dam column}

\item{founder_val}{the value in the sire and dam columns that represent founder}

\item{focal_founder}{the founder for which to calculate the partial kinship matrix}
}
\value{
a partial kinship matrix for the focal founder
}
\description{
Calculate partial kinship matrix for a given pedigree and focal founder
}
