% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_pop_size.R
\name{calc_ped_Ne}
\alias{calc_ped_Ne}
\title{Calculate effective population size from a pedigree}
\usage{
calc_ped_Ne(
  ped,
  Ne = c("Ne_f", "Ne_c"),
  id = "all",
  coanc_mat = NULL,
  inbr = NULL,
  eqg = NULL,
  max_compar = 1e+05
)
}
\arguments{
\item{ped}{A pedigree dataframe in the format output by the process_ped function.}

\item{Ne}{The estimators under which to calculate Ne. Currently, two options are implemented---one based on inbreeding (Ne_f) and another based on coancestry (Ne_c)}

\item{id}{The individuals to be considered in the Ne calculation. The option "all" (the default) uses all individuals in the pedigree in the calculation.}

\item{coanc_mat}{The coancestry (kinship) matrix for the individuals in the pedigree. If this is not supplied, the coancestry matrix will be calculated internally.}

\item{inbr}{A dataframe containing information on inbreeding. The dataframe must include a column named "id" with id of each individual and a column named "f_ped," which records the inbreeding value for each individual. If this is not supplied, the inbreeding information will be calculated internally.}

\item{eqg}{A dataframe containing information on effective number of generations for each individual. The dataframe must include a column named "id" with id of each individual and a column named "equiv_gen," which records the equivalent number of generations for each individual. If this is not supplied, the number of equivalent generations will be calculated internally.}

\item{max_compar}{The number coancestry values to be considered in the calculation of Ne. If the number of values is higher than max_compar, then the coancestry values will be randomly subsampled down to max_compar. This argument is only considered for the "Ne_c" estimator of Ne.}
}
\value{
a list containing the Ne estimates and standard errors
}
\description{
Calculate effective population size from a pedigree
}
