% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_quantities.R
\name{calc_pedmat}
\alias{calc_pedmat}
\title{Calculate kinship/coancestry matrix or additive relationship matrix from a pedigree}
\usage{
calc_pedmat(
  ped,
  id_col = 1,
  sire_col = 2,
  dam_col = 3,
  type = c("kinship", "add_rel")
)
}
\arguments{
\item{ped}{the pedigree from which to calculate the kinship matrix}

\item{id_col}{the name or index of the id column}

\item{sire_col}{the name or index of the sire column}

\item{dam_col}{the name or index of the dam column}

\item{type}{Whether the kinship matrix or additive relationship matrix should be calculated}
}
\value{
the kinship or additive relationship matrix
}
\description{
Calculate kinship/coancestry matrix or additive relationship matrix from a pedigree
}
