% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{character2numeric_id}
\alias{character2numeric_id}
\title{Switch character IDs to numeric (for the ID, sire, and dam columns)}
\usage{
character2numeric_id(ped, id_col, sire_col, dam_col, founder_val = 0)
}
\arguments{
\item{ped}{pedigree}

\item{id_col}{name or index of id column}

\item{sire_col}{name or index of sire column}

\item{dam_col}{name or index of dam column}

\item{founder_val}{the value that represents a founder in the sire and dam cols}
}
\value{
a pedigree stored in a dataframe with updated individual ID values
}
\description{
Switch character IDs to numeric (for the ID, sire, and dam columns)
}
