% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{check_order_ped}
\alias{check_order_ped}
\title{Evaluate whether the pedigree is ordered. NOTE: this has not been thoroughly vetted.}
\usage{
check_order_ped(ped, id_col, sire_col, dam_col)
}
\arguments{
\item{ped}{a ped}

\item{id_col}{name or index of id column}

\item{sire_col}{name or index of sire column}

\item{dam_col}{name or index of dam column}
}
\value{
A logical value indicating whether or not the pedigree is ordered with offspring after parents.
}
\description{
Evaluate whether the pedigree is ordered. NOTE: this has not been thoroughly vetted.
}
