% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pedigree.R
\name{extract_ped}
\alias{extract_ped}
\title{Title}
\usage{
extract_ped(
  ped_sim_output,
  remove_nonbreeding_founders = TRUE,
  pedtools_format = TRUE
)
}
\arguments{
\item{pedtools_format}{}
}
\description{
Title
}
