% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_quantities.R
\name{fped_gdrop}
\alias{fped_gdrop}
\title{Calculation of inbreeding from gene dropping}
\usage{
fped_gdrop(gdrop_output)
}
\arguments{
\item{gdrop_output}{output from the simple_gene_drop function (specifically, the geno_origin dataframe)}
}
\value{
a dataframe containing the inbreeding values
}
\description{
Calculation of inbreeding from gene dropping
}
