% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbr_coanc.R
\name{fped_gdrop_mat}
\alias{fped_gdrop_mat}
\title{Calculate inbreeding from gene dropping (the gene_drop_matrix function)}
\usage{
fped_gdrop_mat(gdrop_mat_output)
}
\arguments{
\item{gdrop_mat_output}{gene dropping output from the gene_drop_matrix function}
}
\value{
a dataframe containing the estimated inbreeding for each individual
}
\description{
Calculate inbreeding from gene dropping (the gene_drop_matrix function)
}
