% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_locus_genedrop.R
\name{gene_drop_matrix}
\alias{gene_drop_matrix}
\title{Single locus gene drop on a pedigree (using matrices to record the alleles)}
\usage{
gene_drop_matrix(ped, sims = 10, report_progress = TRUE)
}
\arguments{
\item{ped}{pedigree (stored in dataframe) with the following organization: column 1 --> id, column 2 --> sire id, column 3 --> dam. Founder parents should be coded as 0s.}

\item{sims}{the number of gene drops to perform}

\item{report_progress}{should the simulation progress be reported?}
}
\value{
a list containing the following items: matrix of alleles derived from the sire (rows = individuals, columns = sims); matrix of alleles derived from the dam (rows = individuals, columns = sims); reorded pedigree; indexed version of reorded pedigree
}
\description{
Single locus gene drop on a pedigree (using matrices to record the alleles)
}
