% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Extract ancestors for one or more individuals}
\usage{
get_ancestors(ped, indiv_vec, include_repeats = TRUE)
}
\arguments{
\item{ped}{a pedigree object}

\item{indiv_vec}{the individuals that for which to extract ancestors}

\item{include_repeats}{if individuals are duplicate ancestors, should duplicates be removed?}
}
\value{
a vector of ancestors
}
\description{
Extract ancestors for one or more individuals
}
