% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_founders}
\alias{get_founders}
\title{Extract the founders from a pedigree}
\usage{
get_founders(ped, founder_vals = "0")
}
\arguments{
\item{ped}{a pedigree object}

\item{founder_vals}{the founder values}
}
\value{
a list with the founders and partial founders (the individuals with only one known parent)
}
\description{
Extract the founders from a pedigree
}
