% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{index_pedigree}
\alias{index_pedigree}
\title{Create indexed version of the pedigree}
\usage{
index_pedigree(ped)
}
\arguments{
\item{ped}{pedigree (stored in a dataframe) with the following organization: column 1 --> id, column 2 --> sire id, column3 --> dam id. Founders should have parents coded as 0.}
}
\value{
A dataframe containing the pedigree with ids transformed to the index values of each individual (and founder parents coded as 0s).
}
\description{
Create indexed version of the pedigree
}
