% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_quantities.R
\name{kinship_gdrop}
\alias{kinship_gdrop}
\title{Calculate coancestry/kinship from gene dropping (the gene_drop_matrix function). NOTE: THIS FUNCTION HAS NOT BEEN TESTED}
\usage{
kinship_gdrop(gdrop_output, samples = "all", show_progress = TRUE)
}
\arguments{
\item{gdrop_output}{output from the simple_gene_drop function (specifically, the geno_origin dataframe)}

\item{samples}{the samples for which you want to calculate kinship values. If "all" is specified, all individuals in the pedigree will be included in the calculations.}

\item{show_progress}{specifies whether the function should report the progress of the calculations}
}
\value{
a dataframe with all pairwise combos of individuals and their associated kinship value
}
\description{
Calculate coancestry/kinship from gene dropping (the gene_drop_matrix function). NOTE: THIS FUNCTION HAS NOT BEEN TESTED
}
