% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbr_coanc.R
\name{partial_founder_fped_gdrop}
\alias{partial_founder_fped_gdrop}
\title{Calculation of partial inbreeding from gene dropping}
\usage{
partial_founder_fped_gdrop(gdrop_output)
}
\arguments{
\item{gdrop_output}{output from the simple_gene_drop function (specifically, the geno_origin dataframe)}
}
\value{
a dataframe containing each individual's inbreeding values contributed by each founder
}
\description{
Calculation of partial inbreeding from gene dropping
}
