% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbr_coanc.R
\name{partial_founder_fped_gdrop_mat}
\alias{partial_founder_fped_gdrop_mat}
\title{Calculate the partial founder inbreeding values from gene dropping (the gene_drop_matrix function)}
\usage{
partial_founder_fped_gdrop_mat(gdrop_mat_output)
}
\arguments{
\item{gdrop_mat_output}{gene dropping output from the gene_drop_matrix function}
}
\value{
a dataframe containing the partial founder inbreeding values (partial_founder_fped) and the proportion of inbreeding that can be attributed to each founder (fped_prop)
}
\description{
Calculate the partial founder inbreeding values from gene dropping (the gene_drop_matrix function)
}
