% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_quantities.R
\name{partial_inbreeding}
\alias{partial_inbreeding}
\title{Calculate the partial inbreeding value for each individual derived from each founder}
\usage{
partial_inbreeding(ped, id_col = 1, sire_col = 2, dam_col = 3, founder_val = 0)
}
\arguments{
\item{ped}{the pedigree from which to calculate the partial kinship matrix}

\item{id_col}{the name or index of the id column}

\item{sire_col}{the name or index of the sire column}

\item{dam_col}{the name or index of the dam column}

\item{founder_val}{the value in the sire and dam columns that represent founder}
}
\value{
a list that contains a list of partial coancestry matrices and a dataframe of partial inbreeding values
}
\description{
Calculate the partial inbreeding value for each individual derived from each founder
}
