% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{ped_summary_stats}
\alias{ped_summary_stats}
\title{Calculate completeness and depth measures of a pedigree}
\usage{
ped_summary_stats(
  ped,
  gen = 1,
  summary = c("total_gen", "equiv_gen", "complete_gen", "mean_compl", "compl_index")
)
}
\arguments{
\item{ped}{A pedigree dataframe in the format output by the process_ped function.}

\item{gen}{The number of generations to consider for the compl_index calculation}

\item{summary}{The measures used to summarize the pedigree}
}
\value{
A dataframe with the completeness and/or depth measures of the pedigree
}
\description{
Calculate completeness and depth measures of a pedigree
}
