% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pedigree.R
\name{pedsim2relateStats}
\alias{pedsim2relateStats}
\title{Convert pedigree dataframe generated from sim_ped_repeat_attempt to pedsim2relateStats format and optionally export}
\usage{
pedsim2relateStats(
  ped_output,
  export = FALSE,
  export_dir = getwd(),
  export_file_name = "relateStats_file.tsv"
)
}
\arguments{
\item{ped_output}{pedigree dataframe in format generated by sim_ped_repeat_attempt}

\item{export}{logical specifying whether the formatted pedigree should be exported}

\item{export_dir}{directory for where the file should be exported (if export_dir = TRUE)}

\item{export_file_name}{name for the exported file (if export_dir = TRUE)}
}
\value{
a dataframe of pedigree information in relateStats format
}
\description{
Convert pedigree dataframe generated from sim_ped_repeat_attempt to pedsim2relateStats format and optionally export
}
