% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedutils-package.R
\docType{package}
\name{pedutils-package}
\alias{pedutils}
\alias{pedutils-package}
\title{pedutils: Utility package for working with pedigrees}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alex Lewanski \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
