% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.plotBackend.R
\name{plot_backend}
\alias{plot_backend}
\title{plotting backend modified from kinship2}
\usage{
plot_backend(
  x,
  annot,
  feature.name = NULL,
  which.proband = NULL,
  add.ids = x$id[x$add == 1],
  id = x$id,
  status = x$status,
  affected = x$affected,
  cex = 1,
  col = 1,
  symbolsize = 1,
  branch = 0.6,
  packed = TRUE,
  align = c(1.5, 2),
  width = 10,
  density = c(-1, 35, 65, 20),
  main_title = "Example Pedigree",
  mar = c(4.1, 1, 4.1, 1),
  angle = c(90, 65, 40, 0),
  keep.par = FALSE,
  subregion,
  pconnect = 0.5,
  ...
)
}
\arguments{
\item{x}{a pedigree dataframe}

\item{annot}{the annotation string}

\item{feature.name}{the names of the features}
}
\description{
plotting backend modified from kinship2
}
