% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PlotPed.R
\name{plotPed}
\alias{plotPed}
\title{a convient wrapper for creating plot for pedigree data}
\usage{
plotPed(
  ped,
  annot.cancers = "all",
  annot.features = NULL,
  cex = 0.8,
  branch = 1,
  ...
)
}
\arguments{
\item{ped}{a pedigree that contains \code{ID, Sex, MotherID, FatherID, isProband, CurAge}}

\item{annot.cancers}{the cancers shortnames to display. When set to default /code{NULL}, the top four cancers will be displayed.}

\item{annot.feature}{ONE feature that we would get annotation from pedigree, one of the choices from \code{c("ethic","Twins","CurAge","race")}}
}
\description{
a convient wrapper for creating plot for pedigree data
}
