% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_processing.R
\name{process_ped}
\alias{process_ped}
\title{Process a pedigree}
\usage{
process_ped(
  ped,
  id_col,
  sire_col,
  dam_col,
  sex_col,
  founder_val,
  sex_vals,
  keep_extra_cols = TRUE,
  disable_sex_check = FALSE
)
}
\arguments{
\item{ped}{The unprocessed dataframe. It needs to contain the following columns: id_col (the ID of each individual), sire_col (the ID of the sire), dam_col (the ID of the dam), and sex_col (the sex of the individual).}

\item{id_col}{The name of the id column.}

\item{sire_col}{The name of the sire (father) column.}

\item{dam_col}{The name of the dam (mother) column.}

\item{sex_col}{The name of the sex column.}

\item{founder_val}{One or more (stored as a vector) values that represent the value for founder parents.}

\item{sex_vals}{A list containing the values representing each sex in sex_col. The list should be in the following format: list(male = c('male_val1', male_val2'), female = 'female_val', unknown = c('u1', 'u2', NA)).}

\item{keep_extra_cols}{Logical value (TRUE or FALSE) indicating whether to keep extra columns in the pedigree beyond the id, sire, dam, and sex columns.}

\item{disable_sex_check}{Logical value (TRUE or FALSE) indicating whether sex information should be evaluated.}
}
\value{
A processed pedigree stored as a dataframe.
}
\description{
Process a pedigree
}
