% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_quantities.R
\name{quantify_ancestry}
\alias{quantify_ancestry}
\title{Quantify founder ancestry from gene dropping (the gene_drop_matrix function)}
\usage{
quantify_ancestry(gdrop_mat_output, founder_group_info)
}
\arguments{
\item{gdrop_mat_output}{gene dropping output from the gene_drop_matrix function}

\item{founder_group_info}{the group membership of each founder. The dataframe should have one column named "id" that includes the founder ID and another column named "group" that indicates group membership}
}
\value{
a dataframe containing the ancestry proportion from each group
}
\description{
Quantify founder ancestry from gene dropping (the gene_drop_matrix function)
}
