% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{reorder_ped}
\alias{reorder_ped}
\title{Reorder pedigree so that all offspring follow their parents}
\usage{
reorder_ped(ped, id_col, sire_col, dam_col)
}
\arguments{
\item{ped}{a pedigree}

\item{id_col}{name or index of id column}

\item{sire_col}{name or index of sire column}

\item{dam_col}{name or index of dam column}
}
\value{
a pedigree ordered so that offspring follow their parents
}
\description{
Reorder pedigree so that all offspring follow their parents
}
