% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.buildGenoMat.R
\name{sim.allele2Geno}
\alias{sim.allele2Geno}
\title{Convert a Person's Allele Matrix to a Vector of Genotypes}
\usage{
sim.allele2Geno(alleles)
}
\arguments{
\item{alleles}{allele matrix (\code{2} by \code{number of genes})}
}
\description{
Uses Mendelian laws to convert a person's alleles for a set of genes into
their corresponding genotypes. Returns a vector of zeros and ones of the
same length as \code{number of genes}. A 0 genotype corresponds to two 0
alleles  and a 1 genotype corresponds to at least one 1 allele.
}
\seealso{
Other simulations: 
\code{\link{sim.buildBranchOfAlleleMats}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.inheritAlleleMat}()},
\code{\link{sim.linkParents}()},
\code{\link{sim.simAlleleMat}()},
\code{\link{sim.simCancerVars}()},
\code{\link{sim.simCurAgeVar}()}
}
\concept{simulations}
