% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.buildGenoMat.R
\name{sim.buildAlleleArray}
\alias{sim.buildAlleleArray}
\title{Build the Allele Array for All Family Members in a 4-Generation Family}
\usage{
sim.buildAlleleArray(
  alleleFreq,
  nChildPatern,
  nChildMatern,
  nChild,
  nGrandchildInBranches
)
}
\arguments{
\item{alleleFreq}{allele frequencies for each gene of interest}

\item{nChildPatern}{vector of length \code{2}, indicating the number of daughters
(father's sisters) and number of sons (father + his brothers) in the paternal branch}

\item{nChildMatern}{vector of length \code{2}, indicating the number of daughters
(mother + her sisters) and number of sons (mother's brothers) in the maternal branch}

\item{nChild}{vector of length \code{2}, indicating the number of daughters
(proband + her sisters) and number of sons (proband's brothers) in the proband's
generation}

\item{nGrandchildInBranches}{\code{2} by \code{sum(nChild)} matrix, indicating
the number of daughters and number of sons that the proband and each of her siblings
have. Each row corresponds to the children of one of the members of \code{nChild},
with the proband as the first row.}
}
\description{
Returns a 3D array of alleles, with dimensions \code{number of members}
by \code{2} by \code{number of genes}
}
\seealso{
Other simulations: 
\code{\link{sim.buildBranchOfAlleleMats}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.inheritAlleleMat}()},
\code{\link{sim.linkParents}()},
\code{\link{sim.simAlleleMat}()},
\code{\link{sim.simCancerVars}()},
\code{\link{sim.simCurAgeVar}()},
\code{\link{sim.simFam}()}
}
\concept{simulations}
