% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.buildBranchOfAlleleMats.R
\name{sim.buildBranchOfAlleleMats}
\alias{sim.buildBranchOfAlleleMats}
\title{Build the Allele Matrices for the Progeny of a Branch}
\usage{
sim.buildBranchOfAlleleMats(alleleFreq, nProg, alleles, allelesSpouse)
}
\arguments{
\item{alleleFreq}{allele frequencies for each gene of interest}

\item{nProg}{number of progeny in this branch}

\item{alleles}{\code{2} by \code{number of genes} matrix of alleles
for family member who creates this branch (optional)}

\item{allelesSpouse}{\code{2} by \code{number of genes} matrix of alleles
for family member's spouse (optional)}
}
\description{
Returns a 3d array of allele matrices with dimensions for the progeny (and
possibly one or more of the parents) of the branch. The first dimension of the 3d
array will be equal to the number of progeny plus up to two additional parents.
The other two dimensions are \code{2} by \code{number of genes}. See Details.
}
\details{
If \code{alleles} and/or \code{allelesSpouse} are not supplied, they will
be simulated. In this case, the simulated allele matrices will be added along the
first dimension of the 3d array, i.e. the first dimension will grow to
\code{nProg+1} or \code{nProg+2}.
}
\seealso{
Other simulations: 
\code{\link{sim.buildAlleleArray}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.inheritAlleleMat}()},
\code{\link{sim.linkParents}()},
\code{\link{sim.simAlleleMat}()},
\code{\link{sim.simCancerVars}()},
\code{\link{sim.simCurAgeVar}()},
\code{\link{sim.simFam}()}
}
\concept{simulations}
