% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.buildBranchOfAlleleMats.R
\name{sim.inheritAlleleMat}
\alias{sim.inheritAlleleMat}
\title{Simulate a Person's Matrix of Alleles for a Series of Genes, Based on Inheritance}
\usage{
sim.inheritAlleleMat(alleles1, alleles2)
}
\arguments{
\item{alleles1}{\code{2} by \code{number of genes} allele matrix for parent 1}

\item{alleles2}{\code{2} by \code{number of genes} allele matrix for parent 2}
}
\description{
Simulate a person's alleles for a series of genes, based on their parents' alleles
and Mendelian laws of inheritance. Returns a \code{2} by \code{number of genes}
matrix of zeros and ones, where the first row corresponds to the alleles inherited
from the father and the second row corresponds to the alleles inherited from the
mother.
}
\seealso{
Other simulations: 
\code{\link{sim.allele2Geno}()},
\code{\link{sim.buildBranchOfAlleleMats}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.linkParents}()},
\code{\link{sim.simAlleleMat}()},
\code{\link{sim.simCancerVars}()},
\code{\link{sim.simCurAgeVar}()}
}
\concept{simulations}
