% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.linkParents.R
\name{sim.linkParents}
\alias{sim.linkParents}
\title{Generate Unique IDs and Sex in a Branch and Link Parents' IDs}
\usage{
sim.linkParents(
  nDaughters,
  nSons,
  mothID,
  fathID,
  lastID = 0,
  includeParents = TRUE
)
}
\arguments{
\item{nDaughters}{number of daughters in branch}

\item{nSons}{number of sons in branch}

\item{mothID}{mother's ID, if known (optional)}

\item{fathID}{father's ID, if known (optional)}

\item{lastID}{the largest ID number that is already in use.}

\item{includeParents}{boolean flag, indicating whether to assign IDs to the
parents. Defaults to TRUE
Defaults to 0 (i.e. new IDs will start from 1)}
}
\description{
Returns a data frame with 4 columns: \code{ID}, \code{Sex}, \code{MotherID},
and \code{FatherID}. There will be a row for each of the progeny, plus a row for
any parent that were simulated/not blood a relation of the main family, i.e.
spouses or the grandparents. \code{ID} stores the individuals' ID, \code{Sex}
stores their sex, and \code{MotherID} and \code{FatherID} stores their
parent's IDs. A parental ID of 0 corresponds to a parent who is not part of the
family matrix.
}
\details{
If the \code{mothID} and/or \code{fathID} is not known, the mother
and/or father is assumed to have been simulated.
}
\seealso{
Other simulations: 
\code{\link{sim.buildAlleleArray}()},
\code{\link{sim.buildBranchOfAlleleMats}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.inheritAlleleMat}()},
\code{\link{sim.simAlleleMat}()},
\code{\link{sim.simCancerVars}()},
\code{\link{sim.simCurAgeVar}()},
\code{\link{sim.simFam}()}
}
\concept{simulations}
