% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.buildBranchOfAlleleMats.R
\name{sim.simAlleleMat}
\alias{sim.simAlleleMat}
\title{Simulate a Person's Matrix of Alleles for a Series of Genes}
\usage{
sim.simAlleleMat(alleleFreq)
}
\arguments{
\item{alleleFreq}{allele frequencies for each gene of interest}
}
\description{
Returns a \code{2} by \code{number of genes} matrix of zeros and ones,
indicating which alleles the person has for each of the \code{length(alleleFreq)}
genes, simulated based on the supplied allele frequencies.
}
\seealso{
Other simulations: 
\code{\link{sim.allele2Geno}()},
\code{\link{sim.buildBranchOfAlleleMats}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.inheritAlleleMat}()},
\code{\link{sim.linkParents}()},
\code{\link{sim.simCancerVars}()},
\code{\link{sim.simCurAgeVar}()}
}
\concept{simulations}
