% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.simCancerVars.R
\name{sim.simCancerVars}
\alias{sim.simCancerVars}
\title{Generate the \code{isAff} and \code{Age} Variables for Each Cancer}
\usage{
sim.simCancerVars(
  genoMat,
  Gender,
  CurAge,
  CP,
  PG,
  cancers,
  ageMax = 94,
  ageMin = 2,
  censoring = TRUE,
  affTime = FALSE
)
}
\arguments{
\item{genoMat}{genotype matrix for the family (one row for each family member).}

\item{Gender}{indicator vector (1 for males, 0 for females) of same length as
number of rows in \code{genoMat}.}

\item{CurAge}{vector of current ages for each family member; of same length as
number of rows in \code{genoMat}.}

\item{CP}{list of cancer penetrance matrices, one for each gender.}

\item{PG}{vector of possible genotypes.}

\item{cancers}{vector of cancers to include in the model.}

\item{ageMax}{maximum age to consider. Defaults to 94.}

\item{ageMin}{minimum age to consider. Defaults to 2.}

\item{censoring}{if FALSE, then will assume everyone without any cancers lives
to the maximum age. Defaults to TRUE.}

\item{affTime}{if TRUE, the output will include the cancer times (which may
or may not be observed due to censoring). Defaults to FALSE}
}
\description{
Generates the \code{isAff} indicator and \code{Age} of diagnosis variables for
each cancer, plus \code{isAffAny} (indicator for having any cancer) and
\code{AgeAny} (age of diagnosis for first cancer). Also returns an overall
\code{isDead} indicator variable and an updated \code{CurAge} variable based
on the minimum of the death age and original current age. Incorporates
censoring. Returns matrix with dimensions \code{number of members} by
\code{2 x (number of cancers) + 4}.
}
\details{
Assumes naming conventions for cancers are consistent between arguments.
The first column in the output matrix will be the current age \code{CurAge} and
last column will be the indicator for death \code{isDead}. The
\code{2 x (number of cancers) + 2} columns in the middle are the \code{isAff}
and \code{Age} variables for each cancer, plus the two columns for any cancer.
}
\seealso{
Other simulations: 
\code{\link{sim.allele2Geno}()},
\code{\link{sim.buildBranchOfAlleleMats}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.inheritAlleleMat}()},
\code{\link{sim.linkParents}()},
\code{\link{sim.simAlleleMat}()},
\code{\link{sim.simCurAgeVar}()}
}
\concept{simulations}
