% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.simCurAgeVar.R
\name{sim.simCurAgeVar}
\alias{sim.simCurAgeVar}
\title{Simulate Current Ages for Branch}
\usage{
sim.simCurAgeVar(nProg, mothAge, fathAge)
}
\arguments{
\item{nProg}{number of progeny in branch}

\item{mothAge}{mother's age, if known (optional)}

\item{fathAge}{father's age, if known (optional)}
}
\description{
Returns a numeric vector of current ages for the progeny of the
branch, plus any parents who have missing ages (parents who were simulated/not
blood relations of the main family, i.e. spouses or the grandparents). The
childrens' ages are simulated based on their mother's age.
}
\details{
Does not ensure that the children have different ages from each other
(i.e., two siblings with the same ages are not necessarily twins).
If one of the parents' ages is missing, it will be simulated based on the other
parent's age. If both are missing, they will be assumed to be at the top of tree
(grandparents). The grandmother's age will be simulated from a Norm(100, 2)
distribution and the grandfather's age will be simulated based on her age.
}
\seealso{
Other simulations: 
\code{\link{sim.buildAlleleArray}()},
\code{\link{sim.buildBranchOfAlleleMats}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.inheritAlleleMat}()},
\code{\link{sim.linkParents}()},
\code{\link{sim.simAlleleMat}()},
\code{\link{sim.simCancerVars}()},
\code{\link{sim.simFam}()}
}
\concept{simulations}
