% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.simFam.R
\name{sim.simFam}
\alias{sim.simFam}
\title{Underlying Function to Simulate a Family/Pedigree Matrix}
\usage{
sim.simFam(
  nSibsPatern,
  nSibsMatern,
  nSibs,
  nGrandchild,
  alleleFreq,
  CP,
  genes,
  cancers,
  maxMut = 2,
  ageMax = 94,
  ageMin = 2,
  includeGeno = FALSE,
  includeGrandparents = TRUE,
  censoring = TRUE,
  genderPro = NULL,
  genoMat = NULL,
  CurAge = NULL,
  affTime = FALSE,
  BiomarkerTesting = NULL,
  includeBiomarkers = FALSE,
  maxTries = 5
)
}
\arguments{
\item{nSibsPatern}{vector of length \code{2}, indicating the number of sisters
and brothers the father has (does not include father)}

\item{nSibsMatern}{vector of length \code{2}, indicating the number of sisters
and brothers the mother has (does not include mother)}

\item{nSibs}{vector of length \code{2}, indicating the number of sisters and
brothers the proband has (does not include proband)}

\item{nGrandchild}{\code{sum(nSibs)+1} by \code{2} matrix, indicating the number
of daughters and number of sons that the proband and each of her siblings have.
Each row corresponds to the children of one of the members of \code{nChild}, with
the proband as the first row. Alternatively, \code{nGrandchild} can be passed in
as a vector of length \code{2}, indicating the number of daughters and number of
sons that the proband and each of her siblings have, when they each have the same
number.}

\item{alleleFreq}{allele frequencies for each gene of interest (named vector).}

\item{CP}{list of cancer penetrance matrices, separated by each gender.}

\item{genes}{vector of genes to include in the model.}

\item{cancers}{vector of cancers to include in the model.}

\item{maxMut}{maximum number of mutations to consider. Defaults to 2.}

\item{ageMax}{maximum age to consider. Defaults to 94.}

\item{ageMin}{minimum age to consider. Defaults to 2.}

\item{includeGeno}{boolean flag, indicating whether or not to include genotype
matrix in output. Defaults to FALSE, and was mostly used for troubleshooting.}

\item{includeGrandparents}{boolean flag, indicating whether to drop the
grandparents from the final pedigree. Defaults to TRUE.}

\item{censoring}{if FALSE, then will assume that nobody dies and that we
observe everybody's cancer outcomes until they reach \code{ageMax}. Defaults to TRUE.}

\item{genderPro}{Can be set to "Female" or "Male" to designate the gender of the proband.
If left as NULL, the proband's gender will be randomly generated.}

\item{genoMat}{genotype matrix for the family. This is a matrix where the number of rows is
the number of family members and the number of columns is the number of mutations. If
left as NULL (the default), the genotypes will be generated.}

\item{CurAge}{vector of current ages for the family. If left as NULL (the default),
the current ages will be generated.}

\item{affTime}{if TRUE, the output will include the cancer times (which may
or may not be observed due to censoring). Defaults to FALSE}

\item{BiomarkerTesting}{list of biomarker testing sensitivities and specificities.}

\item{includeBiomarkers}{boolean flag, indicating whether or not to simulate biomarkers
(Breast and Colorectal currently supported). Defaults to FALSE.}

\item{maxTries}{number of runs to attempt for simulating the allele
array and genotype matrix for all family members. Pathogenic variants
on both alleles for a given gene is currently not supported. When this
occurs or an individual has a simulated genotype with more than
maxMut simultaneous mutations, the function's solution is to re-simulate
the alleles/genotypes up to maxTries times. If an admissible allele array
or genotype matrix cannot be simulated after maxTries attempts, an
error will be raised. These scenarios are more likely to occur for large
values of alleleFreq. Defaults to 5 (each).}
}
\description{
Returns a completed family matrix with the following columns for each member:
\itemize{
\item \code{ID} = Member identifier
\item \code{MotherID} = Mother's identifier number
\item \code{FatherID} = Father's identifier number
\item \code{Sex} = 1 for males, 0 for females
\item \code{isProband} = Indicator for whether or not person is proband
\item \code{CurAge} = Family member's current age, or age at death if dead
\item \code{isAffCancer} variables = Indicators for whether or not person
had cancer type \code{Cancer}, e.g. \code{isAffBC}, \code{isAffO}, etc, plus
an indicator \code{isAffAny} for having any cancer
\item \code{AgeCancer} variables = Age of cancer diagnosis for cancer type
\code{Cancer}, e.g. \code{isAffBC}, \code{isAffO}, etc, plus the age of
first cancer diagnosis \code{AgeAny}. 0 if person did not have that cancer
type
\item \code{isDead} = Indicator for whether or not person is dead
\item \code{Twins} = Twin status, set to 0 for not twins (not simulated)
\item \code{Ancestry} = Ancestry, set to "nonAJ" (not simulated)
\item \code{race} = Race, set to "All_Races" (not simulated)
\item \code{riskmod} = Risk modifier/intervention types, set to empty for
no risk modifiers (not simulated)
\item \code{interAge} = Age of risk modifiers/interventions, set to
empty for no risk modifiers (not simulated)
}
}
\details{
Assumes naming conventions for cancers are consistent between arguments.
}
\seealso{
Other simulations: 
\code{\link{sim.buildAlleleArray}()},
\code{\link{sim.buildBranchOfAlleleMats}()},
\code{\link{sim.buildGenoMat}()},
\code{\link{sim.inheritAlleleMat}()},
\code{\link{sim.linkParents}()},
\code{\link{sim.simAlleleMat}()},
\code{\link{sim.simCancerVars}()},
\code{\link{sim.simCurAgeVar}()}
}
\concept{simulations}
