% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pedigree.R
\name{sim_ped_repeat_attempt}
\alias{sim_ped_repeat_attempt}
\title{Simulate a simple population and record pedigree}
\usage{
sim_ped_repeat_attempt(
  attempts = 10L,
  founders = 20L,
  cycles = 5L,
  start_year = 2020,
  age_mort_horizontal_shift = -1.5,
  age_mort_slope = 0.425,
  max_hatchling_survive_prob = 0.8,
  prob_pair = 0.8,
  pair_up_order = c("f_to_m", "m_to_f"),
  min_age_reproduction = 1,
  add_immigrants = c("custom", "random", "none"),
  immigrant_vec = NULL,
  immigrant_age = 1,
  immigrant_prob = 2,
  founder_age_lambda = 4,
  offspring_lambda = 5,
  safeguard = 5000,
  max_pop = 1000,
  report_progress = TRUE
)
}
\arguments{
\item{attempts}{If the simulation "fails" (the population goes extinct before the number of cycles), attempts is the number of times to retry the simulation to achieve success.}

\item{founders}{Specifications for the founders of the population. The user can either input the founders or have the function generate founders by inputting an integer, which defines the the number of founders to create. If the user wants to input the founders, the dataframe needs to include the following columns in the specified order: 'id', 'sex', 'birth_year', 'age'}

\item{cycles}{The number of years/reproduction cycles (accepts an integer value; default = 5).}

\item{start_year}{The starting year of the population (accepts an integer value; default 2020).}

\item{age_mort_horizontal_shift}{This is equivalent to the beta0 term in a logistic regression. age_mort_horizontal_shift shifts the horizontal location of the transition from low to high probability (or vice versa) with smaller values shifting the curve to the right (default = -1.5).}

\item{age_mort_slope}{This is equivalent to the beta1 term in a logistic regression. age_mort_slope alters the slope of function with higher values resulting in a more severe transition from low to high probability (or vice versa).}

\item{max_hatchling_survive_prob}{Maximum probability of hatchling survival. This is the survival probability when the population size is 0 (numeric value between 0 and 1; default = 0.8  numeric between 0 and 1).}

\item{prob_pair}{Probability of mating each year for the sex that is first chosen for mating (currently age independent)}

\item{pair_up_order}{The order in which the sexes are paired up. 'f_to_m' chooses females first and then pairs them (with replacement) with males. 'm_to_f' does the opposite}

\item{min_age_reproduction}{The minimum age threshold for possible reproduction. For example, an age of 1 means that all individuals can potentially reproduce the year after they are born}

\item{add_immigrants}{How and whether migrants should be added to the population. There are 3 options: (1) none: no migrants are added (2) random: for each cycle, the number of migrants is drawn from a poisson distribution parameterized with founder_age_lambda (3) custom: the user provides the number of migrants for each cycle using the immigrant_vec argument.}

\item{immigrant_vec}{A vector containing the number of migrants for each cycle. The vector needs to be the same length as the number of cycles. This argument is only relevant when add_immigrants is set to custom.}

\item{immigrant_age}{What age should the immigrants be considered (this could be made more flexible)}

\item{immigrant_prob}{The poisson lambda value for determine the number of migrants to add each cycle.}

\item{founder_age_lambda}{The poisson lambda value for determining age of founders when add_immigrants is set to random.}

\item{offspring_lambda}{The poisson value for the number of offspring for each breeding pair.}

\item{safeguard}{If the population size exceeds this value, the simulation will be halted this is included to prevent a population explosion. NOTE: this is largely superfluous now because of population regulation via max_pop.}

\item{max_pop}{The population carrying capacity, which is used to determine offspring survival. This is used for population size regulation.}

\item{report_progress}{Specifies whether the function should report the progress of the simulation.}
}
\value{
A dataframe containing the population simulation output.
}
\description{
Simulate a simple population and record pedigree
}
