% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_locus_genedrop.R
\name{simple_gene_drop}
\alias{simple_gene_drop}
\title{Single locus gene drop on a pedigree}
\usage{
simple_gene_drop(
  ped,
  id_col,
  sire_col,
  dam_col,
  sex_col,
  sims = 10,
  pop_freq = 0.5,
  fixed_founder_genotypes = FALSE,
  calc_contribution = FALSE,
  report_progress = TRUE
)
}
\arguments{
\item{ped}{pedigree on which to perform gene drop}

\item{id_col}{name or index of id column}

\item{sire_col}{name or index of sire column}

\item{dam_col}{name or index of dam column}

\item{sex_col}{name or index of sex column}

\item{sims}{number of gene drop simulations}

\item{pop_freq}{the frequency of the allele in the starting population}

\item{fixed_founder_genotypes}{should founder genotypes be fixed?}

\item{calc_contribution}{should the ancestry contributions of each founder be recorded?}

\item{report_progress}{should the simulation progress be shown?}
}
\value{
a list of simulation output
}
\description{
Single locus gene drop on a pedigree
}
