% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_locus_genedrop.R
\name{single_gene_drop}
\alias{single_gene_drop}
\title{Title}
\usage{
single_gene_drop(
  ped_size,
  pedigree,
  sire_draw_vec,
  dam_draw_vec,
  geno_mat,
  geno_orig_mat,
  f_vec_mat
)
}
\arguments{
\item{f_vec_mat}{}
}
\description{
Title
}
