% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_locus_genedrop.R
\name{single_genetic_contr}
\alias{single_genetic_contr}
\title{Title}
\usage{
single_genetic_contr(
  ped_size,
  pedigree,
  sire_draw_vec,
  dam_draw_vec,
  contr_mat
)
}
\arguments{
\item{contr_mat}{}
}
\description{
Title
}
