% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{subset_ped_descendants}
\alias{subset_ped_descendants}
\title{Subset pedigree down to a focal set of individuals and their descendants}
\usage{
subset_ped_descendants(ped, indivs)
}
\arguments{
\item{ped}{pedigree (stored in dataframe) with the following organization: column 1 --> id, column 2 --> sire id, column 3 --> dam. Founder parents should be coded as 0s.}

\item{indivs}{the individuals for whom you want to subset the pedigree}
}
\value{
a dataframe containing pedigree information for the focal individuals and their descendants based on the input pedigree
}
\description{
Subset pedigree down to a focal set of individuals and their descendants
}
