% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pedigree.R
\name{survival_explanatory_vis}
\alias{survival_explanatory_vis}
\title{Plot survival probability}
\usage{
survival_explanatory_vis(
  age_mort_horizontal_shift = -1.5,
  age_mort_slope = 0.6,
  min_plotting_age = 0,
  max_plotting_age = 20,
  max_pop = 1000,
  max_hatchling_survive_prob = 0.6
)
}
\arguments{
\item{age_mort_horizontal_shift}{This is equivalent to the beta0 term in a logistic regression. age_mort_horizontal_shift shifts the horizontal location of the transition from low to high probability (or vice versa) with smaller values shifting the curve to the right (default = -1.5).}

\item{age_mort_slope}{This is equivalent to the beta1 term in a logistic regression. age_mort_slope alters the slope of function with higher values resulting in a more severe transition from low to high probability (or vice versa).}

\item{min_plotting_age}{The minimum age to include in the plot.}

\item{max_plotting_age}{The maximum age to include in the plot.}

\item{max_pop}{The population carrying capacity, which is used to determine offspring survival. This is used for population size regulation.}

\item{max_hatchling_survive_prob}{Maximum probability of hatchling survival. This is the survival probability when the population size is 0.}
}
\value{
A plot (ggplot object) visualizing the relationship between age and survival probability.
}
\description{
Plot survival probability
}
