% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pedigree.R
\name{extract_ped}
\alias{extract_ped}
\title{Extract pedigree from population simulation}
\usage{
extract_ped(
  ped_sim_output,
  remove_nonbreeding_founders = TRUE,
  pedtools_format = TRUE
)
}
\arguments{
\item{ped_sim_output}{Population simulation output from the sim_ped_repeat_attempt function.}

\item{remove_nonbreeding_founders}{Logical (TRUE or FALSE) indicating whether to remove founders from the population that didn't have any offspring.}

\item{pedtools_format}{Logical (TRUE or FALSE) indicating whether to output the pedigree in pedtools format.}
}
\value{
A dataframe containing pedigree information.
}
\description{
Extract pedigree from population simulation
}
