% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_locus_genedrop.R
\name{simple_gene_drop}
\alias{simple_gene_drop}
\title{Title}
\usage{
simple_gene_drop(
  ped,
  id_col,
  sire_col,
  dam_col,
  sex_col,
  sims = 10,
  pop_freq = 0.5,
  fixed_founder_genotypes = FALSE,
  calc_contribution = FALSE,
  report_progress = TRUE
)
}
\arguments{
\item{report_progress}{}
}
\description{
Title
}
