% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pedigree.R
\name{survival_explanatory_vis}
\alias{survival_explanatory_vis}
\title{Plot survival probability}
\usage{
survival_explanatory_vis(
  age_mort_horizontal_shift = -1.5,
  age_mort_slope = 0.6,
  min_plotting_age = 0,
  max_plotting_age = 20,
  max_pop = 1000,
  max_hatchling_survive_prob = 0.6
)
}
\arguments{
\item{max_hatchling_survive_prob}{}
}
\description{
Plot survival probability
}
